+++ src/stop.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 set -e
-# shellcheck source=../lib/container.sh
-source $YODA_PATH/lib/container.sh
+# shellcheck disable=SC1091 source=../lib/container.sh
+source "$YODA_PATH/lib/container.sh"
 containers=$(get_containers "$@")
 
-$YODA_CMD compose > $COMPOSE_FILE
-docker-compose stop -t $STOP_WAIT_TIMEOUT $containers
+$YODA_CMD compose > "$COMPOSE_FILE"
+docker-compose stop -t "$STOP_WAIT_TIMEOUT" $containers